Move Mesh::ReadGmshMesh to separate file
