+++ mesh/mesh_readers.cpp
@@ -10,12 +10,12 @@
 // CONTRIBUTING.md for details.
 
 #include "mesh_headers.hpp"
+#include "gmsh.hpp"
 #include "ncnurbs.hpp"
 #include "../fem/fem.hpp"
 #include "../general/binaryio.hpp"
 #include "../general/text.hpp"
 #include "../general/tinyxml2.h"
-#include "gmsh.hpp"
 
 #include <iostream>
 #include <cstdio>
@@ -1512,1338 +1512,7 @@ void Mesh::ReadInlineMesh(std::istream &input, bool generate_edges)
    }
 }
 
-void Mesh::ReadGmshMesh(std::istream &input, int &curved, int &read_gf)
-{
-   string buff;
-   real_t version;
-   int binary, dsize;
-   input >> version >> binary >> dsize;
-   if (version < 2.2)
-   {
-      MFEM_ABORT("Gmsh file version < 2.2");
-   }
-   if (dsize != sizeof(double))
-   {
-      MFEM_ABORT("Gmsh file : dsize != sizeof(double)");
-   }
-   getline(input, buff);
-   // There is a number 1 in binary format
-   if (binary)
-   {
-      int one;
-      input.read(reinterpret_cast<char*>(&one), sizeof(one));
-      if (one != 1)
-      {
-         MFEM_ABORT("Gmsh file : wrong binary format");
-      }
-   }
-
-   // A map between a serial number of the vertex and its number in the file
-   // (there may be gaps in the numbering, and also Gmsh enumerates vertices
-   // starting from 1, not 0)
-   map<int, int> vertices_map;
-
-   // A map containing names of physical curves, surfaces, and volumes.
-   // The first index is the dimension of the physical manifold, the second
-   // index is the element attribute number of the set, and the string is
-   // the assigned name.
-   map<int,map<int,std::string> > phys_names_by_dim;
-
-   // Gmsh always outputs coordinates in 3D, but MFEM distinguishes between the
-   // mesh element dimension (Dim) and the dimension of the space in which the
-   // mesh is embedded (spaceDim). For example, a 2D MFEM mesh has Dim = 2 and
-   // spaceDim = 2, while a 2D surface mesh in 3D has Dim = 2 but spaceDim = 3.
-   // Below we set spaceDim by measuring the mesh bounding box and checking for
-   // a lower dimensional subspace. The assumption is that the mesh is at least
-   // 2D if the y-dimension of the box is non-trivial and 3D if the z-dimension
-   // is non-trivial. Note that with these assumptions a 2D mesh parallel to the
-   // yz plane will be considered a surface mesh embedded in 3D whereas the same
-   // 2D mesh parallel to the xy plane will be considered a 2D mesh.
-   real_t bb_tol = 1e-14;
-   real_t bb_min[3];
-   real_t bb_max[3];
-
-   // Mesh order
-   int mesh_order = 1;
-
-   // Mesh type
-   bool periodic = false;
-
-   // Vector field to store uniformly spaced Gmsh high order mesh coords
-   GridFunction Nodes_gf;
-
-   // Read the lines of the mesh file. If we face specific keyword, we'll treat
-   // the section.
-   while (input >> buff)
-   {
-      if (buff == "$Nodes") // reading mesh vertices
-      {
-         input >> NumOfVertices;
-         getline(input, buff);
-         vertices.SetSize(NumOfVertices);
-         int serial_number;
-         const int gmsh_dim = 3; // Gmsh always outputs 3 coordinates
-         real_t coord[gmsh_dim];
-         for (int ver = 0; ver < NumOfVertices; ++ver)
-         {
-            if (binary)
-            {
-               input.read(reinterpret_cast<char*>(&serial_number), sizeof(int));
-               input.read(reinterpret_cast<char*>(coord), gmsh_dim*sizeof(double));
-            }
-            else // ASCII
-            {
-               input >> serial_number;
-               for (int ci = 0; ci < gmsh_dim; ++ci)
-               {
-                  input >> coord[ci];
-               }
-            }
-            vertices[ver] = Vertex(coord, gmsh_dim);
-            vertices_map[serial_number] = ver;
-
-            for (int ci = 0; ci < gmsh_dim; ++ci)
-            {
-               bb_min[ci] = (ver == 0) ? coord[ci] :
-                            std::min(bb_min[ci], coord[ci]);
-               bb_max[ci] = (ver == 0) ? coord[ci] :
-                            std::max(bb_max[ci], coord[ci]);
-            }
-         }
-         real_t bb_size = std::max(bb_max[0] - bb_min[0],
-                                   std::max(bb_max[1] - bb_min[1],
-                                            bb_max[2] - bb_min[2]));
-         spaceDim = 1;
-         if (bb_max[1] - bb_min[1] > bb_size * bb_tol)
-         {
-            spaceDim++;
-         }
-         if (bb_max[2] - bb_min[2] > bb_size * bb_tol)
-         {
-            spaceDim++;
-         }
-
-         if (static_cast<int>(vertices_map.size()) != NumOfVertices)
-         {
-            MFEM_ABORT("Gmsh file : vertices indices are not unique");
-         }
-      } // section '$Nodes'
-      else if (buff == "$Elements") // reading mesh elements
-      {
-         int num_of_all_elements;
-         input >> num_of_all_elements;
-         // = NumOfElements + NumOfBdrElements + (maybe, PhysicalPoints)
-         getline(input, buff);
-
-         int serial_number; // serial number of an element
-         int type_of_element; // ID describing a type of a mesh element
-         int n_tags; // number of different tags describing an element
-         int phys_domain; // element's attribute
-         int elem_domain; // another element's attribute (rarely used)
-         int n_partitions; // number of partitions where an element takes place
-
-         // number of nodes for each type of Gmsh elements, type is the index of
-         // the array + 1
-         int nodes_of_gmsh_element[] =
-         {
-            2,   // 2-node line.
-            3,   // 3-node triangle.
-            4,   // 4-node quadrangle.
-            4,   // 4-node tetrahedron.
-            8,   // 8-node hexahedron.
-            6,   // 6-node prism.
-            5,   // 5-node pyramid.
-            3,   /* 3-node second order line (2 nodes associated with the
-                    vertices and 1 with the edge). */
-            6,   /* 6-node second order triangle (3 nodes associated with the
-                    vertices and 3 with the edges). */
-            9,   /* 9-node second order quadrangle (4 nodes associated with the
-                    vertices, 4 with the edges and 1 with the face). */
-            10,  /* 10-node second order tetrahedron (4 nodes associated with
-                    the vertices and 6 with the edges). */
-            27,  /* 27-node second order hexahedron (8 nodes associated with the
-                    vertices, 12 with the edges, 6 with the faces and 1 with the
-                    volume). */
-            18,  /* 18-node second order prism (6 nodes associated with the
-                    vertices, 9 with the edges and 3 with the quadrangular
-                    faces). */
-            14,  /* 14-node second order pyramid (5 nodes associated with the
-                    vertices, 8 with the edges and 1 with the quadrangular
-                    face). */
-            1,   // 1-node point.
-            8,   /* 8-node second order quadrangle (4 nodes associated with the
-                    vertices and 4 with the edges). */
-            20,  /* 20-node second order hexahedron (8 nodes associated with the
-                    vertices and 12 with the edges). */
-            15,  /* 15-node second order prism (6 nodes associated with the
-                    vertices and 9 with the edges). */
-            13,  /* 13-node second order pyramid (5 nodes associated with the
-                    vertices and 8 with the edges). */
-            9,   /* 9-node third order incomplete triangle (3 nodes associated
-                    with the vertices, 6 with the edges) */
-            10,  /* 10-node third order triangle (3 nodes associated with the
-                    vertices, 6 with the edges, 1 with the face) */
-            12,  /* 12-node fourth order incomplete triangle (3 nodes associated
-                    with the vertices, 9 with the edges) */
-            15,  /* 15-node fourth order triangle (3 nodes associated with the
-                    vertices, 9 with the edges, 3 with the face) */
-            15,  /* 15-node fifth order incomplete triangle (3 nodes associated
-                    with the vertices, 12 with the edges) */
-            21,  /* 21-node fifth order complete triangle (3 nodes associated
-                    with the vertices, 12 with the edges, 6 with the face) */
-            4,   /* 4-node third order edge (2 nodes associated with the
-                    vertices, 2 internal to the edge) */
-            5,   /* 5-node fourth order edge (2 nodes associated with the
-                    vertices, 3 internal to the edge) */
-            6,   /* 6-node fifth order edge (2 nodes associated with the
-                    vertices, 4 internal to the edge) */
-            20,  /* 20-node third order tetrahedron (4 nodes associated with the
-                    vertices, 12 with the edges, 4 with the faces) */
-            35,  /* 35-node fourth order tetrahedron (4 nodes associated with
-                    the vertices, 18 with the edges, 12 with the faces, and 1
-                    with the volume) */
-            56,  /* 56-node fifth order tetrahedron (4 nodes associated with the
-                    vertices, 24 with the edges, 24 with the faces, and 4 with
-                    the volume) */
-            -1,-1, /* unsupported tetrahedral types */
-            -1,-1, /* unsupported polygonal and polyhedral types */
-            16,  /* 16-node third order quadrilateral (4 nodes associated with
-                    the vertices, 8 with the edges, 4 with the face) */
-            25,  /* 25-node fourth order quadrilateral (4 nodes associated with
-                    the vertices, 12 with the edges, 9 with the face) */
-            36,  /* 36-node fifth order quadrilateral (4 nodes associated with
-                    the vertices, 16 with the edges, 16 with the face) */
-            -1,-1,-1, /* unsupported quadrilateral types */
-            28,  /* 28-node sixth order complete triangle (3 nodes associated
-                    with the vertices, 15 with the edges, 10 with the face) */
-            36,  /* 36-node seventh order complete triangle (3 nodes associated
-                    with the vertices, 18 with the edges, 15 with the face) */
-            45,  /* 45-node eighth order complete triangle (3 nodes associated
-                    with the vertices, 21 with the edges, 21 with the face) */
-            55,  /* 55-node ninth order complete triangle (3 nodes associated
-                    with the vertices, 24 with the edges, 28 with the face) */
-            66,  /* 66-node tenth order complete triangle (3 nodes associated
-                    with the vertices, 27 with the edges, 36 with the face) */
-            49,  /* 49-node sixth order quadrilateral (4 nodes associated with
-                    the vertices, 20 with the edges, 25 with the face) */
-            64,  /* 64-node seventh order quadrilateral (4 nodes associated with
-                    the vertices, 24 with the edges, 36 with the face) */
-            81,  /* 81-node eighth order quadrilateral (4 nodes associated with
-                    the vertices, 28 with the edges, 49 with the face) */
-            100, /* 100-node ninth order quadrilateral (4 nodes associated with
-                    the vertices, 32 with the edges, 64 with the face) */
-            121, /* 121-node tenth order quadrilateral (4 nodes associated with
-                    the vertices, 36 with the edges, 81 with the face) */
-            -1,-1,-1,-1,-1, /* unsupported triangular types */
-            -1,-1,-1,-1,-1, /* unsupported quadrilateral types */
-            7,   /* 7-node sixth order edge (2 nodes associated with the
-                    vertices, 5 internal to the edge) */
-            8,   /* 8-node seventh order edge (2 nodes associated with the
-                    vertices, 6 internal to the edge) */
-            9,   /* 9-node eighth order edge (2 nodes associated with the
-                    vertices, 7 internal to the edge) */
-            10,  /* 10-node ninth order edge (2 nodes associated with the
-                    vertices, 8 internal to the edge) */
-            11,  /* 11-node tenth order edge (2 nodes associated with the
-                    vertices, 9 internal to the edge) */
-            -1,  /* unsupported linear types */
-            -1,-1,-1, /* unsupported types */
-            84,  /* 84-node sixth order tetrahedron (4 nodes associated with the
-                    vertices, 30 with the edges, 40 with the faces, and 10 with
-                    the volume) */
-            120, /* 120-node seventh order tetrahedron (4 nodes associated with
-                    the vertices, 36 with the edges, 60 with the faces, and 20
-                    with the volume) */
-            165, /* 165-node eighth order tetrahedron (4 nodes associated with
-                    the vertices, 42 with the edges, 84 with the faces, and 35
-                    with the volume) */
-            220, /* 220-node ninth order tetrahedron (4 nodes associated with
-                    the vertices, 48 with the edges, 112 with the faces, and 56
-                    with the volume) */
-            286, /* 286-node tenth order tetrahedron (4 nodes associated with
-                    the vertices, 54 with the edges, 144 with the faces, and 84
-                    with the volume) */
-            -1,-1,-1,       /* undefined types */
-            -1,-1,-1,-1,-1, /* unsupported tetrahedral types */
-            -1,-1,-1,-1,-1,-1, /* unsupported types */
-            40,  /* 40-node third order prism (6 nodes associated with the
-                    vertices, 18 with the edges, 14 with the faces, and 2 with
-                    the volume) */
-            75,  /* 75-node fourth order prism (6 nodes associated with the
-                    vertices, 27 with the edges, 33 with the faces, and 9 with
-                    the volume) */
-            64,  /* 64-node third order hexahedron (8 nodes associated with the
-                    vertices, 24 with the edges, 24 with the faces and 8 with
-                    the volume).*/
-            125, /* 125-node fourth order hexahedron (8 nodes associated with
-                    the vertices, 36 with the edges, 54 with the faces and 27
-                    with the volume).*/
-            216, /* 216-node fifth order hexahedron (8 nodes associated with the
-                    vertices, 48 with the edges, 96 with the faces and 64 with
-                    the volume).*/
-            343, /* 343-node sixth order hexahedron (8 nodes associated with the
-                    vertices, 60 with the edges, 150 with the faces and 125 with
-                    the volume).*/
-            512, /* 512-node seventh order hexahedron (8 nodes associated with
-                    the vertices, 72 with the edges, 216 with the faces and 216
-                    with the volume).*/
-            729, /* 729-node eighth order hexahedron (8 nodes associated with
-                    the vertices, 84 with the edges, 294 with the faces and 343
-                    with the volume).*/
-            1000,/* 1000-node ninth order hexahedron (8 nodes associated with
-                    the vertices, 96 with the edges, 384 with the faces and 512
-                    with the volume).*/
-            -1,-1,-1,-1,-1,-1,-1, /* unsupported hexahedron types */
-            126, /* 126-node fifth order prism (6 nodes associated with the
-                    vertices, 36 with the edges, 60 with the faces, and 24 with
-                    the volume) */
-            196, /* 196-node sixth order prism (6 nodes associated with the
-                    vertices, 45 with the edges, 95 with the faces, and 50 with
-                    the volume) */
-            288, /* 288-node seventh order prism (6 nodes associated with the
-                    vertices, 54 with the edges, 138 with the faces, and 90 with
-                    the volume) */
-            405, /* 405-node eighth order prism (6 nodes associated with the
-                    vertices, 63 with the edges, 189 with the faces, and 147
-                    with the volume) */
-            550, /* 550-node ninth order prism (6 nodes associated with the
-                    vertices, 72 with the edges, 248 with the faces, and 224
-                    with the volume) */
-            -1,-1,-1,-1,-1,-1,-1, /* unsupported prism types */
-            30,  /* 30-node third order pyramid (5 nodes associated with the
-                    vertices, 16 with the edges and 8 with the faces, and 1 with
-                    the volume). */
-            55,  /* 55-node fourth order pyramid (5 nodes associated with the
-                    vertices, 24 with the edges and 21 with the faces, and 5
-                    with the volume). */
-            91,  /* 91-node fifth order pyramid (5 nodes associated with the
-                    vertices, 32 with the edges and 40 with the faces, and 14
-                    with the volume). */
-            140, /* 140-node sixth order pyramid (5 nodes associated with the
-                    vertices, 40 with the edges and 65 with the faces, and 30
-                    with the volume). */
-            204, /* 204-node seventh order pyramid (5 nodes associated with the
-                    vertices, 48 with the edges and 96 with the faces, and 55
-                    with the volume). */
-            285, /* 285-node eighth order pyramid (5 nodes associated with the
-                    vertices, 56 with the edges and 133 with the faces, and 91
-                    with the volume). */
-            385  /* 385-node ninth order pyramid (5 nodes associated with the
-                    vertices, 64 with the edges and 176 with the faces, and 140
-                    with the volume). */
-         };
-
-         /** The following mappings convert the Gmsh node orderings for high
-             order elements to MFEM's L2 degree of freedom ordering. To support
-             more options examine Gmsh's ordering and read off the indices in
-             MFEM's order. For example 2nd order Gmsh quadrilaterals use the
-             following ordering:
-
-                3--6--2
-                |  |  |
-                7  8  5
-                |  |  |
-                0--4--1
-
-             (from https://gmsh.info/doc/texinfo/gmsh.html#Node-ordering)
-
-             Whereas MFEM uses a tensor product ordering with the horizontal
-             axis cycling fastest so we would read off:
-
-                0 4 1 7 8 5 3 6 2
-
-             This corresponds to the quad9 mapping below.
-         */
-         int lin3[] = {0,2,1};                // 2nd order segment
-         int lin4[] = {0,2,3,1};              // 3rd order segment
-         int tri6[] = {0,3,1,5,4,2};          // 2nd order triangle
-         int tri10[] = {0,3,4,1,8,9,5,7,6,2}; // 3rd order triangle
-         int quad9[] = {0,4,1,7,8,5,3,6,2};   // 2nd order quadrilateral
-         int quad16[] = {0,4,5,1,11,12,13,6,  // 3rd order quadrilateral
-                         10,15,14,7,3,9,8,2
-                        };
-         int tet10[] {0,4,1,6,5,2,7,9,8,3};   // 2nd order tetrahedron
-         int tet20[] = {0,4,5,1,9,16,6,8,7,2, // 3rd order tetrahedron
-                        11,17,15,18,19,13,10,14,12,3
-                       };
-         int hex27[] {0,8,1,9,20,11,3,13,2,   // 2nd order hexahedron
-                      10,21,12,22,26,23,15,24,14,
-                      4,16,5,17,25,18,7,19,6
-                     };
-         int hex64[] {0,8,9,1,10,32,35,14,    // 3rd order hexahedron
-                      11,33,34,15,3,19,18,2,
-                      12,36,37,16,40,56,57,44,
-                      43,59,58,45,22,49,48,20,
-                      13,39,38,17,41,60,61,47,
-                      42,63,62,46,23,50,51,21,
-                      4,24,25,5,26,52,53,28,
-                      27,55,54,29,7,31,30,6
-                     };
-
-         int wdg18[] = {0,6,1,7,9,2,8,15,10,    // 2nd order wedge/prism
-                        16,17,11,3,12,4,13,14,5
-                       };
-         int wdg40[] = {0,6,7,1,8,24,12,9,13,2, // 3rd order wedge/prism
-                        10,26,27,14,30,38,34,33,35,16,
-                        11,29,28,15,31,39,37,32,36,17,
-                        3,18,19,4,20,25,22,21,23,5
-                       };
-
-         int pyr14[] = {0,5,1,6,13,8,3,          // 2nd order pyramid
-                        10,2,7,9,12,11,4
-                       };
-         int pyr30[] = {0,5,6,1,7,25,28,11,8,26, // 3rd order pyramid
-                        27,12,3,16,15,2,9,21,13,22,
-                        29,23,19,24,17,10,14,20,18,4
-                       };
-
-         vector<Element*> elements_0D, elements_1D, elements_2D, elements_3D;
-         elements_0D.reserve(num_of_all_elements);
-         elements_1D.reserve(num_of_all_elements);
-         elements_2D.reserve(num_of_all_elements);
-         elements_3D.reserve(num_of_all_elements);
-
-         // Temporary storage for high order vertices, if present
-         vector<Array<int>*> ho_verts_1D, ho_verts_2D, ho_verts_3D;
-         ho_verts_1D.reserve(num_of_all_elements);
-         ho_verts_2D.reserve(num_of_all_elements);
-         ho_verts_3D.reserve(num_of_all_elements);
-
-         // Temporary storage for order of elements
-         vector<int> ho_el_order_1D, ho_el_order_2D, ho_el_order_3D;
-         ho_el_order_1D.reserve(num_of_all_elements);
-         ho_el_order_2D.reserve(num_of_all_elements);
-         ho_el_order_3D.reserve(num_of_all_elements);
-
-         // Vertex order mappings
-         Array<int*> ho_lin(11); ho_lin = NULL;
-         Array<int*> ho_tri(11); ho_tri = NULL;
-         Array<int*> ho_sqr(11); ho_sqr = NULL;
-         Array<int*> ho_tet(11); ho_tet = NULL;
-         Array<int*> ho_hex(10); ho_hex = NULL;
-         Array<int*> ho_wdg(10); ho_wdg = NULL;
-         Array<int*> ho_pyr(10); ho_pyr = NULL;
-
-         // Use predefined arrays at lowest orders (for efficiency)
-         ho_lin[2] = lin3;  ho_lin[3] = lin4;
-         ho_tri[2] = tri6;  ho_tri[3] = tri10;
-         ho_sqr[2] = quad9; ho_sqr[3] = quad16;
-         ho_tet[2] = tet10; ho_tet[3] = tet20;
-         ho_hex[2] = hex27; ho_hex[3] = hex64;
-         ho_wdg[2] = wdg18; ho_wdg[3] = wdg40;
-         ho_pyr[2] = pyr14; ho_pyr[3] = pyr30;
-
-         bool has_nonpositive_phys_domain = false;
-         bool has_positive_phys_domain = false;
-
-         if (binary)
-         {
-            int n_elem_part = 0; // partial sum of elements that are read
-            const int header_size = 3;
-            // header consists of 3 numbers: type of the element, number of
-            // elements of this type, and number of tags
-            int header[header_size];
-            int n_elem_one_type; // number of elements of a specific type
-
-            while (n_elem_part < num_of_all_elements)
-            {
-               input.read(reinterpret_cast<char*>(header),
-                          header_size*sizeof(int));
-               type_of_element = header[0];
-               n_elem_one_type = header[1];
-               n_tags          = header[2];
-
-               n_elem_part += n_elem_one_type;
-
-               const int n_elem_nodes = nodes_of_gmsh_element[type_of_element-1];
-               vector<int> data(1+n_tags+n_elem_nodes);
-               for (int el = 0; el < n_elem_one_type; ++el)
-               {
-                  input.read(reinterpret_cast<char*>(&data[0]),
-                             data.size()*sizeof(int));
-                  int dd = 0; // index for data array
-                  serial_number = data[dd++];
-                  // physical domain - the most important value (to distinguish
-                  // materials with different properties)
-                  phys_domain = (n_tags > 0) ? data[dd++] : 1;
-                  // elementary domain - to distinguish different geometrical
-                  // domains (typically, it's used rarely)
-                  elem_domain = (n_tags > 1) ? data[dd++] : 0;
-                  // the number of tags is bigger than 2 if there are some
-                  // partitions (domain decompositions)
-                  n_partitions = (n_tags > 2) ? data[dd++] : 0;
-                  // we currently just skip the partitions if they exist, and go
-                  // directly to vertices describing the mesh element
-                  vector<int> vert_indices(n_elem_nodes);
-                  for (int vi = 0; vi < n_elem_nodes; ++vi)
-                  {
-                     map<int, int>::const_iterator it =
-                        vertices_map.find(data[1+n_tags+vi]);
-                     if (it == vertices_map.end())
-                     {
-                        MFEM_ABORT("Gmsh file : vertex index doesn't exist");
-                     }
-                     vert_indices[vi] = it->second;
-                  }
-
-                  // Non-positive attributes are not allowed in MFEM. However,
-                  // by default, Gmsh sets the physical domain of all elements
-                  // to zero. In the case that all elements have physical domain
-                  // zero, we will given them attribute 1. If only some elements
-                  // have physical domain zero, we will throw an error.
-                  if (phys_domain <= 0)
-                  {
-                     has_nonpositive_phys_domain = true;
-                     phys_domain = 1;
-                  }
-                  else
-                  {
-                     has_positive_phys_domain = true;
-                  }
-
-                  // initialize the mesh element
-                  int el_order = 11;
-                  switch (type_of_element)
-                  {
-                     case  1: //   2-node line
-                     case  8: //   3-node line (2nd order)
-                     case 26: //   4-node line (3rd order)
-                     case 27: //   5-node line (4th order)
-                     case 28: //   6-node line (5th order)
-                     case 62: //   7-node line (6th order)
-                     case 63: //   8-node line (7th order)
-                     case 64: //   9-node line (8th order)
-                     case 65: //  10-node line (9th order)
-                     case 66: //  11-node line (10th order)
-                     {
-                        elements_1D.push_back(
-                           new Segment(&vert_indices[0], phys_domain));
-                        if (type_of_element != 1)
-                        {
-                           el_order = n_elem_nodes - 1;
-                           Array<int> * hov = new Array<int>;
-                           hov->Append(&vert_indices[0], n_elem_nodes);
-                           ho_verts_1D.push_back(hov);
-                           ho_el_order_1D.push_back(el_order);
-                        }
-                        break;
-                     }
-                     case  2: el_order--; //  3-node triangle
-                     case  9: el_order--; //  6-node triangle (2nd order)
-                     case 21: el_order--; // 10-node triangle (3rd order)
-                     case 23: el_order--; // 15-node triangle (4th order)
-                     case 25: el_order--; // 21-node triangle (5th order)
-                     case 42: el_order--; // 28-node triangle (6th order)
-                     case 43: el_order--; // 36-node triangle (7th order)
-                     case 44: el_order--; // 45-node triangle (8th order)
-                     case 45: el_order--; // 55-node triangle (9th order)
-                     case 46: el_order--; // 66-node triangle (10th order)
-                        {
-                           elements_2D.push_back(
-                              new Triangle(&vert_indices[0], phys_domain));
-                           if (el_order > 1)
-                           {
-                              Array<int> * hov = new Array<int>;
-                              hov->Append(&vert_indices[0], n_elem_nodes);
-                              ho_verts_2D.push_back(hov);
-                              ho_el_order_2D.push_back(el_order);
-                           }
-                           break;
-                        }
-                     case  3: el_order--; //   4-node quadrangle
-                     case 10: el_order--; //   9-node quadrangle (2nd order)
-                     case 36: el_order--; //  16-node quadrangle (3rd order)
-                     case 37: el_order--; //  25-node quadrangle (4th order)
-                     case 38: el_order--; //  36-node quadrangle (5th order)
-                     case 47: el_order--; //  49-node quadrangle (6th order)
-                     case 48: el_order--; //  64-node quadrangle (7th order)
-                     case 49: el_order--; //  81-node quadrangle (8th order)
-                     case 50: el_order--; // 100-node quadrangle (9th order)
-                     case 51: el_order--; // 121-node quadrangle (10th order)
-                        {
-                           elements_2D.push_back(
-                              new Quadrilateral(&vert_indices[0], phys_domain));
-                           if (el_order > 1)
-                           {
-                              Array<int> * hov = new Array<int>;
-                              hov->Append(&vert_indices[0], n_elem_nodes);
-                              ho_verts_2D.push_back(hov);
-                              ho_el_order_2D.push_back(el_order);
-                           }
-                           break;
-                        }
-                     case  4: el_order--; //   4-node tetrahedron
-                     case 11: el_order--; //  10-node tetrahedron (2nd order)
-                     case 29: el_order--; //  20-node tetrahedron (3rd order)
-                     case 30: el_order--; //  35-node tetrahedron (4th order)
-                     case 31: el_order--; //  56-node tetrahedron (5th order)
-                     case 71: el_order--; //  84-node tetrahedron (6th order)
-                     case 72: el_order--; // 120-node tetrahedron (7th order)
-                     case 73: el_order--; // 165-node tetrahedron (8th order)
-                     case 74: el_order--; // 220-node tetrahedron (9th order)
-                     case 75: el_order--; // 286-node tetrahedron (10th order)
-                        {
-#ifdef MFEM_USE_MEMALLOC
-                           elements_3D.push_back(TetMemory.Alloc());
-                           elements_3D.back()->SetVertices(&vert_indices[0]);
-                           elements_3D.back()->SetAttribute(phys_domain);
-#else
-                           elements_3D.push_back(
-                              new Tetrahedron(&vert_indices[0], phys_domain));
-#endif
-                           if (el_order > 1)
-                           {
-                              Array<int> * hov = new Array<int>;
-                              hov->Append(&vert_indices[0], n_elem_nodes);
-                              ho_verts_3D.push_back(hov);
-                              ho_el_order_3D.push_back(el_order);
-                           }
-                           break;
-                        }
-                     case  5: el_order--; //    8-node hexahedron
-                     case 12: el_order--; //   27-node hexahedron (2nd order)
-                     case 92: el_order--; //   64-node hexahedron (3rd order)
-                     case 93: el_order--; //  125-node hexahedron (4th order)
-                     case 94: el_order--; //  216-node hexahedron (5th order)
-                     case 95: el_order--; //  343-node hexahedron (6th order)
-                     case 96: el_order--; //  512-node hexahedron (7th order)
-                     case 97: el_order--; //  729-node hexahedron (8th order)
-                     case 98: el_order--; // 1000-node hexahedron (9th order)
-                        {
-                           el_order--; // Gmsh does not define an order 10 hex
-                           elements_3D.push_back(
-                              new Hexahedron(&vert_indices[0], phys_domain));
-                           if (el_order > 1)
-                           {
-                              Array<int> * hov = new Array<int>;
-                              hov->Append(&vert_indices[0], n_elem_nodes);
-                              ho_verts_3D.push_back(hov);
-                              ho_el_order_3D.push_back(el_order);
-                           }
-                           break;
-                        }
-                     case   6: el_order--; //   6-node wedge
-                     case  13: el_order--; //  18-node wedge (2nd order)
-                     case  90: el_order--; //  40-node wedge (3rd order)
-                     case  91: el_order--; //  75-node wedge (4th order)
-                     case 106: el_order--; // 126-node wedge (5th order)
-                     case 107: el_order--; // 196-node wedge (6th order)
-                     case 108: el_order--; // 288-node wedge (7th order)
-                     case 109: el_order--; // 405-node wedge (8th order)
-                     case 110: el_order--; // 550-node wedge (9th order)
-                        {
-                           el_order--; // Gmsh does not define an order 10 wedge
-                           elements_3D.push_back(
-                              new Wedge(&vert_indices[0], phys_domain));
-                           if (el_order > 1)
-                           {
-                              Array<int> * hov = new Array<int>;
-                              hov->Append(&vert_indices[0], n_elem_nodes);
-                              ho_verts_3D.push_back(hov);
-                              ho_el_order_3D.push_back(el_order);
-                           }
-                           break;
-                        }
-                     case   7: el_order--; //   5-node pyramid
-                     case  14: el_order--; //  14-node pyramid (2nd order)
-                     case 118: el_order--; //  30-node pyramid (3rd order)
-                     case 119: el_order--; //  55-node pyramid (4th order)
-                     case 120: el_order--; //  91-node pyramid (5th order)
-                     case 121: el_order--; // 140-node pyramid (6th order)
-                     case 122: el_order--; // 204-node pyramid (7th order)
-                     case 123: el_order--; // 285-node pyramid (8th order)
-                     case 124: el_order--; // 385-node pyramid (9th order)
-                        {
-                           el_order--; // Gmsh does not define an order 10 pyr
-                           elements_3D.push_back(
-                              new Pyramid(&vert_indices[0], phys_domain));
-                           if (el_order > 1)
-                           {
-                              Array<int> * hov = new Array<int>;
-                              hov->Append(&vert_indices[0], n_elem_nodes);
-                              ho_verts_3D.push_back(hov);
-                              ho_el_order_3D.push_back(el_order);
-                           }
-                           break;
-                        }
-                     case 15: // 1-node point
-                     {
-                        elements_0D.push_back(
-                           new Point(&vert_indices[0], phys_domain));
-                        break;
-                     }
-                     default: // any other element
-                        MFEM_WARNING("Unsupported Gmsh element type.");
-                        break;
-
-                  } // switch (type_of_element)
-               } // el (elements of one type)
-            } // all elements
-         } // if binary
-         else // ASCII
-         {
-            for (int el = 0; el < num_of_all_elements; ++el)
-            {
-               input >> serial_number >> type_of_element >> n_tags;
-               vector<int> data(n_tags);
-               for (int i = 0; i < n_tags; ++i) { input >> data[i]; }
-               // physical domain - the most important value (to distinguish
-               // materials with different properties)
-               phys_domain = (n_tags > 0) ? data[0] : 1;
-               // elementary domain - to distinguish different geometrical
-               // domains (typically, it's used rarely)
-               elem_domain = (n_tags > 1) ? data[1] : 0;
-               // the number of tags is bigger than 2 if there are some
-               // partitions (domain decompositions)
-               n_partitions = (n_tags > 2) ? data[2] : 0;
-               // we currently just skip the partitions if they exist, and go
-               // directly to vertices describing the mesh element
-               const int n_elem_nodes = nodes_of_gmsh_element[type_of_element-1];
-               vector<int> vert_indices(n_elem_nodes);
-               int index;
-               for (int vi = 0; vi < n_elem_nodes; ++vi)
-               {
-                  input >> index;
-                  map<int, int>::const_iterator it = vertices_map.find(index);
-                  if (it == vertices_map.end())
-                  {
-                     MFEM_ABORT("Gmsh file : vertex index doesn't exist");
-                  }
-                  vert_indices[vi] = it->second;
-               }
-
-               // Non-positive attributes are not allowed in MFEM. However,
-               // by default, Gmsh sets the physical domain of all elements
-               // to zero. In the case that all elements have physical domain
-               // zero, we will given them attribute 1. If only some elements
-               // have physical domain zero, we will throw an error.
-               if (phys_domain <= 0)
-               {
-                  has_nonpositive_phys_domain = true;
-                  phys_domain = 1;
-               }
-               else
-               {
-                  has_positive_phys_domain = true;
-               }
-
-               // initialize the mesh element
-               int el_order = 11;
-               switch (type_of_element)
-               {
-                  case  1: //  2-node line
-                  case  8: //  3-node line (2nd order)
-                  case 26: //  4-node line (3rd order)
-                  case 27: //  5-node line (4th order)
-                  case 28: //  6-node line (5th order)
-                  case 62: //  7-node line (6th order)
-                  case 63: //  8-node line (7th order)
-                  case 64: //  9-node line (8th order)
-                  case 65: // 10-node line (9th order)
-                  case 66: // 11-node line (10th order)
-                  {
-                     elements_1D.push_back(
-                        new Segment(&vert_indices[0], phys_domain));
-                     if (type_of_element != 1)
-                     {
-                        Array<int> * hov = new Array<int>;
-                        hov->Append(&vert_indices[0], n_elem_nodes);
-                        ho_verts_1D.push_back(hov);
-                        el_order = n_elem_nodes - 1;
-                        ho_el_order_1D.push_back(el_order);
-                     }
-                     break;
-                  }
-                  case  2: el_order--; //  3-node triangle
-                  case  9: el_order--; //  6-node triangle (2nd order)
-                  case 21: el_order--; // 10-node triangle (3rd order)
-                  case 23: el_order--; // 15-node triangle (4th order)
-                  case 25: el_order--; // 21-node triangle (5th order)
-                  case 42: el_order--; // 28-node triangle (6th order)
-                  case 43: el_order--; // 36-node triangle (7th order)
-                  case 44: el_order--; // 45-node triangle (8th order)
-                  case 45: el_order--; // 55-node triangle (9th order)
-                  case 46: el_order--; // 66-node triangle (10th order)
-                     {
-                        elements_2D.push_back(
-                           new Triangle(&vert_indices[0], phys_domain));
-                        if (el_order > 1)
-                        {
-                           Array<int> * hov = new Array<int>;
-                           hov->Append(&vert_indices[0], n_elem_nodes);
-                           ho_verts_2D.push_back(hov);
-                           ho_el_order_2D.push_back(el_order);
-                        }
-                        break;
-                     }
-                  case  3: el_order--; //   4-node quadrangle
-                  case 10: el_order--; //   9-node quadrangle (2nd order)
-                  case 36: el_order--; //  16-node quadrangle (3rd order)
-                  case 37: el_order--; //  25-node quadrangle (4th order)
-                  case 38: el_order--; //  36-node quadrangle (5th order)
-                  case 47: el_order--; //  49-node quadrangle (6th order)
-                  case 48: el_order--; //  64-node quadrangle (7th order)
-                  case 49: el_order--; //  81-node quadrangle (8th order)
-                  case 50: el_order--; // 100-node quadrangle (9th order)
-                  case 51: el_order--; // 121-node quadrangle (10th order)
-                     {
-                        elements_2D.push_back(
-                           new Quadrilateral(&vert_indices[0], phys_domain));
-                        if (el_order > 1)
-                        {
-                           Array<int> * hov = new Array<int>;
-                           hov->Append(&vert_indices[0], n_elem_nodes);
-                           ho_verts_2D.push_back(hov);
-                           ho_el_order_2D.push_back(el_order);
-                        }
-                        break;
-                     }
-                  case  4: el_order--; //   4-node tetrahedron
-                  case 11: el_order--; //  10-node tetrahedron (2nd order)
-                  case 29: el_order--; //  20-node tetrahedron (3rd order)
-                  case 30: el_order--; //  35-node tetrahedron (4th order)
-                  case 31: el_order--; //  56-node tetrahedron (5th order)
-                  case 71: el_order--; //  84-node tetrahedron (6th order)
-                  case 72: el_order--; // 120-node tetrahedron (7th order)
-                  case 73: el_order--; // 165-node tetrahedron (8th order)
-                  case 74: el_order--; // 220-node tetrahedron (9th order)
-                  case 75: el_order--; // 286-node tetrahedron (10th order)
-                     {
-#ifdef MFEM_USE_MEMALLOC
-                        elements_3D.push_back(TetMemory.Alloc());
-                        elements_3D.back()->SetVertices(&vert_indices[0]);
-                        elements_3D.back()->SetAttribute(phys_domain);
-#else
-                        elements_3D.push_back(
-                           new Tetrahedron(&vert_indices[0], phys_domain));
-#endif
-                        if (el_order > 1)
-                        {
-                           Array<int> * hov = new Array<int>;
-                           hov->Append(&vert_indices[0], n_elem_nodes);
-                           ho_verts_3D.push_back(hov);
-                           ho_el_order_3D.push_back(el_order);
-                        }
-                        break;
-                     }
-                  case  5: el_order--; //    8-node hexahedron
-                  case 12: el_order--; //   27-node hexahedron (2nd order)
-                  case 92: el_order--; //   64-node hexahedron (3rd order)
-                  case 93: el_order--; //  125-node hexahedron (4th order)
-                  case 94: el_order--; //  216-node hexahedron (5th order)
-                  case 95: el_order--; //  343-node hexahedron (6th order)
-                  case 96: el_order--; //  512-node hexahedron (7th order)
-                  case 97: el_order--; //  729-node hexahedron (8th order)
-                  case 98: el_order--; // 1000-node hexahedron (9th order)
-                     {
-                        el_order--;
-                        elements_3D.push_back(
-                           new Hexahedron(&vert_indices[0], phys_domain));
-                        if (el_order > 1)
-                        {
-                           Array<int> * hov = new Array<int>;
-                           hov->Append(&vert_indices[0], n_elem_nodes);
-                           ho_verts_3D.push_back(hov);
-                           ho_el_order_3D.push_back(el_order);
-                        }
-                        break;
-                     }
-                  case   6: el_order--; //   6-node wedge
-                  case  13: el_order--; //  18-node wedge (2nd order)
-                  case  90: el_order--; //  40-node wedge (3rd order)
-                  case  91: el_order--; //  75-node wedge (4th order)
-                  case 106: el_order--; // 126-node wedge (5th order)
-                  case 107: el_order--; // 196-node wedge (6th order)
-                  case 108: el_order--; // 288-node wedge (7th order)
-                  case 109: el_order--; // 405-node wedge (8th order)
-                  case 110: el_order--; // 550-node wedge (9th order)
-                     {
-                        el_order--;
-                        elements_3D.push_back(
-                           new Wedge(&vert_indices[0], phys_domain));
-                        if (el_order > 1)
-                        {
-                           Array<int> * hov = new Array<int>;
-                           hov->Append(&vert_indices[0], n_elem_nodes);
-                           ho_verts_3D.push_back(hov);
-                           ho_el_order_3D.push_back(el_order);
-                        }
-                        break;
-                     }
-                  case   7: el_order--; //   5-node pyramid
-                  case  14: el_order--; //  14-node pyramid (2nd order)
-                  case 118: el_order--; //  30-node pyramid (3rd order)
-                  case 119: el_order--; //  55-node pyramid (4th order)
-                  case 120: el_order--; //  91-node pyramid (5th order)
-                  case 121: el_order--; // 140-node pyramid (6th order)
-                  case 122: el_order--; // 204-node pyramid (7th order)
-                  case 123: el_order--; // 285-node pyramid (8th order)
-                  case 124: el_order--; // 385-node pyramid (9th order)
-                     {
-                        el_order--;
-                        elements_3D.push_back(
-                           new Pyramid(&vert_indices[0], phys_domain));
-                        if (el_order > 1)
-                        {
-                           Array<int> * hov = new Array<int>;
-                           hov->Append(&vert_indices[0], n_elem_nodes);
-                           ho_verts_3D.push_back(hov);
-                           ho_el_order_3D.push_back(el_order);
-                        }
-                        break;
-                     }
-                  case 15: // 1-node point
-                  {
-                     elements_0D.push_back(
-                        new Point(&vert_indices[0], phys_domain));
-                     break;
-                  }
-                  default: // any other element
-                     MFEM_WARNING("Unsupported Gmsh element type.");
-                     break;
-
-               } // switch (type_of_element)
-            } // el (all elements)
-         } // if ASCII
-
-         if (has_positive_phys_domain && has_nonpositive_phys_domain)
-         {
-            MFEM_ABORT("Non-positive element attribute in Gmsh mesh!\n"
-                       "By default Gmsh sets element tags (attributes)"
-                       " to '0' but MFEM requires that they be"
-                       " positive integers.\n"
-                       "Use \"Physical Curve\", \"Physical Surface\","
-                       " or \"Physical Volume\" to set tags/attributes"
-                       " for all curves, surfaces, or volumes in your"
-                       " Gmsh geometry to values which are >= 1.");
-         }
-         else if (has_nonpositive_phys_domain)
-         {
-            mfem::out << "\nGmsh reader: all element attributes were zero.\n"
-                      << "MFEM only supports positive element attributes.\n"
-                      << "Setting element attributes to 1.\n\n";
-         }
-
-         if (!elements_3D.empty())
-         {
-            Dim = 3;
-            NumOfElements = static_cast<int>(elements_3D.size());
-            elements.SetSize(NumOfElements);
-            for (int el = 0; el < NumOfElements; ++el)
-            {
-               elements[el] = elements_3D[el];
-            }
-            NumOfBdrElements = static_cast<int>(elements_2D.size());
-            boundary.SetSize(NumOfBdrElements);
-            for (int el = 0; el < NumOfBdrElements; ++el)
-            {
-               boundary[el] = elements_2D[el];
-            }
-            for (size_t el = 0; el < ho_el_order_3D.size(); el++)
-            {
-               mesh_order = max(mesh_order, ho_el_order_3D[el]);
-            }
-            // discard other elements
-            for (size_t el = 0; el < elements_1D.size(); ++el)
-            {
-               delete elements_1D[el];
-            }
-            for (size_t el = 0; el < elements_0D.size(); ++el)
-            {
-               delete elements_0D[el];
-            }
-         }
-         else if (!elements_2D.empty())
-         {
-            Dim = 2;
-            NumOfElements = static_cast<int>(elements_2D.size());
-            elements.SetSize(NumOfElements);
-            for (int el = 0; el < NumOfElements; ++el)
-            {
-               elements[el] = elements_2D[el];
-            }
-            NumOfBdrElements = static_cast<int>(elements_1D.size());
-            boundary.SetSize(NumOfBdrElements);
-            for (int el = 0; el < NumOfBdrElements; ++el)
-            {
-               boundary[el] = elements_1D[el];
-            }
-            for (size_t el = 0; el < ho_el_order_2D.size(); el++)
-            {
-               mesh_order = max(mesh_order, ho_el_order_2D[el]);
-            }
-            // discard other elements
-            for (size_t el = 0; el < elements_0D.size(); ++el)
-            {
-               delete elements_0D[el];
-            }
-         }
-         else if (!elements_1D.empty())
-         {
-            Dim = 1;
-            NumOfElements = static_cast<int>(elements_1D.size());
-            elements.SetSize(NumOfElements);
-            for (int el = 0; el < NumOfElements; ++el)
-            {
-               elements[el] = elements_1D[el];
-            }
-            NumOfBdrElements = static_cast<int>(elements_0D.size());
-            boundary.SetSize(NumOfBdrElements);
-            for (int el = 0; el < NumOfBdrElements; ++el)
-            {
-               boundary[el] = elements_0D[el];
-            }
-            for (size_t el = 0; el < ho_el_order_1D.size(); el++)
-            {
-               mesh_order = max(mesh_order, ho_el_order_1D[el]);
-            }
-         }
-         else
-         {
-            MFEM_ABORT("Gmsh file : no elements found");
-            return;
-         }
-
-         if (mesh_order > 1)
-         {
-            curved = 1;
-            read_gf = 0;
-
-            // initialize mesh_geoms so we can create Nodes FE space below
-            this->SetMeshGen();
-
-            // Generate faces and edges so that we can define
-            // FE space on the mesh
-            this->FinalizeTopology();
-
-            // Construct GridFunction for uniformly spaced high order coords
-            FiniteElementCollection* nfec;
-            FiniteElementSpace* nfes;
-            nfec = new L2_FECollection(mesh_order, Dim,
-                                       BasisType::ClosedUniform);
-            nfes = new FiniteElementSpace(this, nfec, spaceDim,
-                                          Ordering::byVDIM);
-            Nodes_gf.SetSpace(nfes);
-            Nodes_gf.MakeOwner(nfec);
-
-            int o = 0;
-            int el_order = 1;
-            for (int el = 0; el < NumOfElements; el++)
-            {
-               const int * vm = NULL;
-               Array<int> * ho_verts = NULL;
-               switch (GetElementType(el))
-               {
-                  case Element::SEGMENT:
-                     ho_verts = ho_verts_1D[el];
-                     el_order = ho_el_order_1D[el];
-                     if (!ho_lin[el_order])
-                     {
-                        ho_lin[el_order] = new int[ho_verts->Size()];
-                        GmshHOSegmentMapping(el_order, ho_lin[el_order]);
-                     }
-                     vm = ho_lin[el_order];
-                     break;
-                  case Element::TRIANGLE:
-                     ho_verts = ho_verts_2D[el];
-                     el_order = ho_el_order_2D[el];
-                     if (!ho_tri[el_order])
-                     {
-                        ho_tri[el_order] = new int[ho_verts->Size()];
-                        GmshHOTriangleMapping(el_order, ho_tri[el_order]);
-                     }
-                     vm = ho_tri[el_order];
-                     break;
-                  case Element::QUADRILATERAL:
-                     ho_verts = ho_verts_2D[el];
-                     el_order = ho_el_order_2D[el];
-                     if (!ho_sqr[el_order])
-                     {
-                        ho_sqr[el_order] = new int[ho_verts->Size()];
-                        GmshHOQuadrilateralMapping(el_order, ho_sqr[el_order]);
-                     }
-                     vm = ho_sqr[el_order];
-                     break;
-                  case Element::TETRAHEDRON:
-                     ho_verts = ho_verts_3D[el];
-                     el_order = ho_el_order_3D[el];
-                     if (!ho_tet[el_order])
-                     {
-                        ho_tet[el_order] = new int[ho_verts->Size()];
-                        GmshHOTetrahedronMapping(el_order, ho_tet[el_order]);
-                     }
-                     vm = ho_tet[el_order];
-                     break;
-                  case Element::HEXAHEDRON:
-                     ho_verts = ho_verts_3D[el];
-                     el_order = ho_el_order_3D[el];
-                     if (!ho_hex[el_order])
-                     {
-                        ho_hex[el_order] = new int[ho_verts->Size()];
-                        GmshHOHexahedronMapping(el_order, ho_hex[el_order]);
-                     }
-                     vm = ho_hex[el_order];
-                     break;
-                  case Element::WEDGE:
-                     ho_verts = ho_verts_3D[el];
-                     el_order = ho_el_order_3D[el];
-                     if (!ho_wdg[el_order])
-                     {
-                        ho_wdg[el_order] = new int[ho_verts->Size()];
-                        GmshHOWedgeMapping(el_order, ho_wdg[el_order]);
-                     }
-                     vm = ho_wdg[el_order];
-                     break;
-                  case Element::PYRAMID:
-                     ho_verts = ho_verts_3D[el];
-                     el_order = ho_el_order_3D[el];
-                     if (!ho_pyr[el_order])
-                     {
-                        ho_pyr[el_order] = new int[ho_verts->Size()];
-                        GmshHOPyramidMapping(el_order, ho_pyr[el_order]);
-                     }
-                     vm = ho_pyr[el_order];
-                     break;
-                  default: // Any other element type
-                     MFEM_WARNING("Unsupported Gmsh element type.");
-                     break;
-               }
-               int nv = (ho_verts) ? ho_verts->Size() : 0;
-
-               for (int v = 0; v<nv; v++)
-               {
-                  real_t * c = GetVertex((*ho_verts)[vm[v]]);
-                  for (int d=0; d<spaceDim; d++)
-                  {
-                     Nodes_gf(spaceDim * (o + v) + d) = c[d];
-                  }
-               }
-               o += nv;
-            }
-         }
-
-         // Delete any high order element to vertex connectivity
-         for (size_t el=0; el<ho_verts_1D.size(); el++)
-         {
-            delete ho_verts_1D[el];
-         }
-         for (size_t el=0; el<ho_verts_2D.size(); el++)
-         {
-            delete ho_verts_2D[el];
-         }
-         for (size_t el=0; el<ho_verts_3D.size(); el++)
-         {
-            delete ho_verts_3D[el];
-         }
-
-         // Delete dynamically allocated high vertex order mappings
-         for (int ord=4; ord<ho_lin.Size(); ord++)
-         {
-            if (ho_lin[ord] != NULL) { delete [] ho_lin[ord]; }
-         }
-         for (int ord=4; ord<ho_tri.Size(); ord++)
-         {
-            if (ho_tri[ord] != NULL) { delete [] ho_tri[ord]; }
-         }
-         for (int ord=4; ord<ho_sqr.Size(); ord++)
-         {
-            if (ho_sqr[ord] != NULL) { delete [] ho_sqr[ord]; }
-         }
-         for (int ord=4; ord<ho_tet.Size(); ord++)
-         {
-            if (ho_tet[ord] != NULL) { delete [] ho_tet[ord]; }
-         }
-         for (int ord=4; ord<ho_hex.Size(); ord++)
-         {
-            if (ho_hex[ord] != NULL) { delete [] ho_hex[ord]; }
-         }
-         for (int ord=4; ord<ho_wdg.Size(); ord++)
-         {
-            if (ho_wdg[ord] != NULL) { delete [] ho_wdg[ord]; }
-         }
-         for (int ord=4; ord<ho_pyr.Size(); ord++)
-         {
-            if (ho_pyr[ord] != NULL) { delete [] ho_pyr[ord]; }
-         }
-
-         // Suppress warnings (MFEM_CONTRACT_VAR does not work here with nvcc):
-         ++n_partitions;
-         ++elem_domain;
-         MFEM_CONTRACT_VAR(n_partitions);
-         MFEM_CONTRACT_VAR(elem_domain);
-
-      } // section '$Elements'
-      else if (buff == "$PhysicalNames") // Named element sets
-      {
-         int num_names = 0;
-         int mdim,num;
-         string name;
-         input >> num_names;
-         for (int i=0; i < num_names; i++)
-         {
-            input >> mdim >> num;
-            getline(input, name);
-
-            // Trim leading white space
-            while (!name.empty() &&
-                   (*name.begin() == ' ' || *name.begin() == '\t'))
-            { name.erase(0,1);}
-
-            // Trim trailing white space
-            while (!name.empty() &&
-                   (*name.rbegin() == ' ' || *name.rbegin() == '\t' ||
-                    *name.rbegin() == '\n' || *name.rbegin() == '\r'))
-            { name.resize(name.length()-1);}
-
-            // Remove enclosing quotes
-            if ( (*name.begin() == '"' || *name.begin() == '\'') &&
-                 (*name.rbegin() == '"' || *name.rbegin() == '\''))
-            {
-               name = name.substr(1,name.length()-2);
-            }
-
-            phys_names_by_dim[mdim][num] = name;
-         }
-      }
-      else if (buff == "$Periodic") // Reading master/slave node pairs
-      {
-         curved = 1;
-         read_gf = 0;
-         periodic = true;
-
-         Array<int> v2v(NumOfVertices);
-         for (int i = 0; i < v2v.Size(); i++)
-         {
-            v2v[i] = i;
-         }
-         int num_per_ent;
-         int num_nodes;
-         input >> num_per_ent;
-         getline(input, buff); // Read end-of-line
-         for (int i = 0; i < num_per_ent; i++)
-         {
-            getline(input, buff); // Read and ignore entity dimension and tags
-            getline(input, buff); // If affine mapping exist, read and ignore
-            if (!strncmp(buff.c_str(), "Affine", 6))
-            {
-               input >> num_nodes;
-            }
-            else
-            {
-               num_nodes = atoi(buff.c_str());
-            }
-            for (int j=0; j<num_nodes; j++)
-            {
-               int slave, master;
-               input >> slave >> master;
-               v2v[slave - 1] = master - 1;
-            }
-            getline(input, buff); // Read end-of-line
-         }
-
-         // Follow existing long chains of slave->master in v2v array.
-         // Upon completion of this loop, each v2v[slave] will point to a true
-         // master vertex. This algorithm is useful for periodicity defined in
-         // multiple directions.
-         for (int slave = 0; slave < v2v.Size(); slave++)
-         {
-            int master = v2v[slave];
-            if (master != slave)
-            {
-               // This loop will end if it finds a circular dependency.
-               while (v2v[master] != master && master != slave)
-               {
-                  master = v2v[master];
-               }
-               if (master == slave)
-               {
-                  // if master and slave are the same vertex, circular dependency
-                  // exists. We need to fix the problem, we choose slave.
-                  v2v[slave] = slave;
-               }
-               else
-               {
-                  // the long chain has ended on the true master vertex.
-                  v2v[slave] = master;
-               }
-            }
-         }
-
-         // Convert nodes to discontinuous GridFunction (if they aren't already)
-         if (mesh_order == 1)
-         {
-            this->FinalizeTopology();
-            this->SetMeshGen();
-            this->SetCurvature(1, true, spaceDim, Ordering::byVDIM);
-         }
-
-         // Replace "slave" vertex indices in the element connectivity
-         // with their corresponding "master" vertex indices.
-         for (int i = 0; i < this->GetNE(); i++)
-         {
-            Element *el = this->GetElement(i);
-            int *v = el->GetVertices();
-            int nv = el->GetNVertices();
-            for (int j = 0; j < nv; j++)
-            {
-               v[j] = v2v[v[j]];
-            }
-         }
-         // Replace "slave" vertex indices in the boundary element connectivity
-         // with their corresponding "master" vertex indices.
-         for (int i = 0; i < this->GetNBE(); i++)
-         {
-            Element *el = this->GetBdrElement(i);
-            int *v = el->GetVertices();
-            int nv = el->GetNVertices();
-            for (int j = 0; j < nv; j++)
-            {
-               v[j] = v2v[v[j]];
-            }
-         }
-      }
-   } // we reach the end of the file
-
-   // Process set names
-   if (phys_names_by_dim.size() > 0)
-   {
-      // Process boundary attribute set names
-      for (auto const &bdr_attr : phys_names_by_dim[Dim-1])
-      {
-         if (!bdr_attribute_sets.AttributeSetExists(bdr_attr.second))
-         {
-            bdr_attribute_sets.CreateAttributeSet(bdr_attr.second);
-         }
-         bdr_attribute_sets.AddToAttributeSet(bdr_attr.second, bdr_attr.first);
-      }
-
-      // Process element attribute set names
-      for (auto const &attr : phys_names_by_dim[Dim])
-      {
-         if (!attribute_sets.AttributeSetExists(attr.second))
-         {
-            attribute_sets.CreateAttributeSet(attr.second);
-         }
-         attribute_sets.AddToAttributeSet(attr.second, attr.first);
-      }
-   }
-
-   this->RemoveUnusedVertices();
-   this->FinalizeTopology();
-
-   // If a high order coordinate field was created project it onto the mesh
-   if (mesh_order > 1)
-   {
-      SetCurvature(mesh_order, periodic, spaceDim, Ordering::byVDIM);
-
-      VectorGridFunctionCoefficient NodesCoef(&Nodes_gf);
-      Nodes->ProjectCoefficient(NodesCoef);
-   }
-}
-
-
 #ifdef MFEM_USE_NETCDF
-
 
 namespace cubit
 {
